+++ packages/core/scripts/src/packagePublisher/index.ts
@@ -1,2 +1,2 @@
 export { getNextRCVersion, publishToNpm } from "./core/npm-utils/npm-utils";
-export { RCPublisher } from "./publish-rc";
+export { RCPublisher } from "./rc-publisher";