+++ packages/core/scripts/src/packagePublisher/publish-rc.types.ts
@@ -1,6 +1,0 @@
-interface PackageInfo {
-  name: string;
-  version: string;
-}
-
-export type { PackageInfo };