+++ packages/core/scripts/CHANGELOG.md
@@ -6,7 +6,7 @@ This package is intended for internal use in generating custom script builds at 
 
 #### 🎉 New features
 
-- Added `publish-rc` script to publish Release Candidate versions of Nimbus packages. ([#300](https://github.com/TiendaNube/nimbus-design-system/pull/300) by [@joacotornello](https://github.com/joacotornello))
+- Added `publish-rc` script to publish Release Candidate versions of Nimbus packages. ([#328](https://github.com/TiendaNube/nimbus-design-system/pull/328) by [@joacotornello](https://github.com/joacotornello))
 
 ## 2025-06-04 `1.7.0`